refactor(telegram): split helpers.ts into helper modules
helpers.ts was 732 lines of pure rendering helpers. Split into a
helpers/ folder:
  - format      — interval parse, duration/token/bytes formatting,
                  model-label/option helpers
  - diagnostics — metrics + doctor HTML rendering
  - menu        — settings/model/backend keyboards, effort rows, the
                  buildModelMenuState builder + UI types

index.ts re-exports everything; telegram + test importers updated to the
barrel path (discord's own helpers.ts untouched). Verified:
telegram-helpers + model-menu + doctor suites pass (47 tests).
