+++ src/__tests__/telegram-model-menu.test.ts
@@ -21,7 +21,7 @@ import {
   renderModelMenuText,
   renderModelBrowseKeyboard,
   type ModelMenuState,
-} from "../frontend/telegram/helpers.js";
+} from "../frontend/telegram/helpers/index.js";
 
 // ── parseModelCallback ─────────────────────────────────────────────────────
