+++ src/frontend/telegram/admin.ts
@@ -17,7 +17,7 @@ import {
 import { getActiveCount } from "../../core/engine/dispatcher.js";
 import { getPulseStatus } from "../../core/background/pulse.js";
 import { getHealthStatus, getRecentErrors } from "../../util/watchdog.js";
-import { formatDuration, formatModelLabel } from "./helpers.js";
+import { formatDuration, formatModelLabel } from "./helpers/index.js";
 
 export async function handleAdminCommand(
   ctx: Context,
